Migrate last test to mock-fetch
