+++ __tests__/projects.mock.js
@@ -119,3 +119,111 @@ describe('RepositoryProjectsManager.assing() posts requests to the API', () => {
     expect(outputTitles).toEqual(titles);
   });
 });
+
+describe('RepositoryProjectsManager.assing() posts requests to the API', () => {
+  let rpm;
+
+  beforeEach(() => {
+    const apiWrapper = new ApiWrapper({ octokit: null });
+
+    jest.spyOn(apiWrapper, 'fetchRepositoryAndProjects')
+      .mockImplementation(() => ({
+        repository: {
+          name: 'example-repository',
+          id: 'R_0000000001',
+          projectsV2: {
+            nodes: [
+              {
+                id: 'PVT_0000000000000001',
+                title: 'layer-100/bar',
+                number: 1099,
+                fields: {
+                  nodes: [
+                    {},
+                    {},
+                    {
+                      id: 'PVTSSF_00000000000000000000001',
+                      name: 'Status',
+                      options: [
+                        {
+                          id: '00000001',
+                          name: 'Todo',
+                        },
+                        {
+                          id: '00000002',
+                          name: 'In Progress',
+                        },
+                        {
+                          id: '00000003',
+                          name: 'Done',
+                        },
+                      ],
+                    },
+                    {},
+                    {},
+                  ],
+                },
+              },
+            ],
+            pageInfo: {
+              hasNextPage: false,
+              endCursor: 'Nw',
+            },
+          },
+        },
+      }));
+
+    jest.spyOn(apiWrapper, 'fetchRepositoryAndProjects')
+      .mockImplementationOnce(() => ({
+        name: 'example-repository',
+        id: 'R_0000000001',
+        projectsV2: {
+          nodes: [
+            {
+              id: 'PVT_000000000000001',
+              title: 'layer-100/bar-1',
+            },
+          ],
+        },
+      }));
+
+    jest.spyOn(apiWrapper, 'fetchItemForPRId')
+      .mockImplementation(() => ({
+        updateProjectV2ItemFieldValue: {
+          projectV2Item: {
+            id: 'PVTI_0000000000000001',
+          },
+        },
+      }));
+
+    jest.spyOn(apiWrapper, 'fetchItemForPRId')
+      .mockImplementation(() => ('PVTI_0000000000000001'));
+
+    jest.spyOn(apiWrapper, 'fetchAssignedProjects')
+      .mockImplementationOnce(() => ([
+        {
+          id: 'PVT_0000000000000001',
+          title: 'layer-100/bar',
+        },
+      ]))
+      .mockImplementationOnce(() => ([]));
+
+    jest.spyOn(apiWrapper, 'deleteProjectItem')
+      .mockImplementation(() => ({
+        deleteProjectV2Item: { deletedItemId: 'PVTI_00000000000000000000000' },
+      }));
+
+    rpm = new RepositoryProjectsManager({ apiWrapper, ownerName: 'acme', repositoryName: 'example-repository' });
+  });
+
+  test('when the PR is assigend to one project but should not be assigned to any project', async () => {
+    const titles = [];
+
+    const pullRequestNumber = 'PR_0000000000000001';
+
+    const outputProjects = await rpm.assign(pullRequestNumber, titles);
+    const outputTitles = outputProjects.map((p) => p.title);
+
+    expect(outputTitles).toEqual(titles);
+  });
+});
